requirements.txt
pandas
duckdb
xgboost
shap
scikit-learn
matplotlib
joblib
pytest
pytest-cov
pre-commit
flask
isort
black
flake8
python-dotenv
beautifulsoup4
html5lib
lxml
requests
psutil
tqdm
pyarrow
selenium
webdriver-manager
pyyaml
